Split Gitify file into 2 variants
One for Romanesco, and one for the derivative project. This is needed to control what gets built the first time, and what gets fetched from upstream after that.

Basically the Romanesco .gitify will only contain Romanesco / Patternlab elements (under the 100K barrier, and NO elements of packages) and the project-specific .gitify will contain everything, so it can be fully deployed on the live server.