Combine Accordion and Tabs CBs
Because their contents are so similar, this was a fairly easy thing to do. Now you can change their layout type without having to transfer all the content to a new CB. This makes it way more usable during prototyping, where you might not be entirely sure yet if you want to use an accordion or tabs.

The trade-off is that some field settings are specific to 1 layout type, but are also visible now when the other one is selected. Fortunately we're only talking a few settings, so the extra clutter is minimal.

What would be really cool for the next iteration, is to have a Tabs block that can scale back into an accordion on smaller screens. Tabs generally suck on mobile, so that would be a big UX improvement. Much like the image gallery CB turning into a slider on mobile. Only issue is that the HTML and JS of an accordion differs quite much from that of the tabs.