Divide pattern layout into multiple columns
The top 2 columns are displayed side-by-side (except for all Electrons). The first column contains a preview of the pattern and can be adjusted in width. The second column contains the source code. Below, there's a full-width third column that contains the pattern info, which is added semi-automatically through a CB template in a Tabs or Accordion field.