Add styling for wireframe grids
Wireframe grids can be used to create small previews of specific layouts using HTML. By giving the columns a light background color, the main areas of a layout can be mimicked in an SUI grid. In the front-end library for example, all templates use this wireframe grid for their preview.