Improve default layout inside HeaderVertical sidebar
Most styling was applied to the #menu-vertical element itself, but that already receives a lot of styling on the attached .menu class. This resulted in conflicting rules and unnecessary overrides.

Especially the different font-size was problematic, since that's also the control setting for em-based sizes inside the element. By applying margins and padding on the #header parent instead, these em sizes can match the rest of the page again.