Add buttons for clearing individual custom cache partitions
The custom partitions are listed as sub items under Manage > Clear Custom Cache. The parent button can also be clicked to clear all partitions at once, after which the page will be refreshed (to work around the 2nd click not firing issue). This behavior is similar to the regular Clear Cache button, which is also positioned right next to it.

The only issue remaining is that the console doesn't parse any feedback at all, but that's too much ExtJS fiddling for what it's worth right now.