Extract settings to different data folder
System and Configuration settings are now built separately as a default / run once config instead. This means that Romanesco Soil only functions as a boilerplate for these settings now, and that they won't be altered anymore on core updates. I hope this will make the rebase process a little easier.