Add Gitify config for building Romanesco
This config contains a lot of internal boundaries, primarily based on the 100K autoincrement threshold. This ensures that only core Romanesco elements will be touched when you run a Gitify build with this config.

Although this config is intended for the build command only, extracts can also be written back to the now separate Romanesco data folder. Core changes inside projects could be committed back to Romanesco this way, but keep in mind that this bypasses the GPM patterns package and thus is not 100% water proof. Could work well though for quick- / hotfixes.